refactor(core): moving this styling wrapper to its own file (as Core.js)
